Homebrew and payload SDK + toolchain updates
## Homebrew applications

### Fonts and text
- Bind the full `libSceFont` surface — all 131 linkable functions plus every struct, enum and font-set id.
- Add `SystemFont` (built-in console fonts, no font file, wide script coverage) and a shared `ScalableFont` base; move `TrueTypeFont` onto it.
- Fix the line-height/baseline scale bug (render and layout scales move together); make measuring and drawing share one kerning-aware advance so they cannot drift; add faux bold/italic (`SetWeight`/`SetSlant`).

### Agc (GPU)
- Sampler: fix the level-of-detail bias range (signed 6.8), give `Create()` the correct defaults, validate anisotropy, add every missing field; truncate fixed-point conversions to match the hardware layer.
- Texture descriptor: add the missing setters (min-LOD clamp, MSAA-depth, border-colour swizzle, DCC) with bounds guards; correct the DCC colour-transform polarity.
- Buffer: fix the structured out-of-bounds default.

### VideoOut, 2D, UI, images
- Fix the five-parameter `ConfigureOutput`/`IsOutputSupported` ABI and the `VblankStatus` layout; bind every remaining header-defined VideoOut function and struct (events, output status, HDR/colour-space, alpha/blend, window margins, flip master).
- Alpha-correct off-screen compositing; clip/truncate text across nine UI controls; transform mesh normals by the inverse-transpose; correct the IME error codes and option flags.
- Image codecs: GIF interlace hang fix plus a new `GifEncoder`; BMP and TGA decoders now read the full range of forms; JPEG applies EXIF orientation on decode and adds 4:2:2/grayscale encode; surface bilinear, span-fill and clamp fixes.

## Payloads

### Fixes

- Fixed four bugs preventing NativeAOT managed code from running standalone (without kstuff-lite loaded first): mmap stack corruption, breadcrumb string embedding, ThreadBlockKeyInit alignment, GCHeapHardLimit 1MB to 16MB typo
- Made all managed-side kernel access firmware-aware using sysctl-based FW detection and existing 57-firmware offset tables — the same binary now works on any supported firmware (1.00 through 12.70)
- Fixed two broken samples (`list-files`, `test-privileges`) referencing non-existent `KernelOffsets1001` type
- Added `Asm.Disp32()` to CompatEmitter for internal RIP-relative fixups in compat functions

### kstuffSharp port

- Now 2-binary architecture: separate installer payload (with ReturnOnExit CRT) embedded into the loader, matching the original ps5-kstuff-ldr + ps5-kstuff/payload.bin design
- Single final payload output
- 40+ fixes across 23 files

+++ src/SharpProspero/Graphics/Surface.Shapes.cs
@@ -158,10 +158,12 @@ public readonly unsafe partial struct Surface
             sweep = MathF.Tau;
         bool fullTurn = sweep >= MathF.Tau - 1e-4f;
 
-        int x0 = Math.Max(0, cx - outerRadius);
-        int y0 = Math.Max(0, cy - outerRadius);
-        int x1 = Math.Min(Width - 1, cx + outerRadius);
-        int y1 = Math.Min(Height - 1, cy + outerRadius);
+        // Combine centre and radius in 64-bit so an extreme centre or radius cannot wrap the scan
+        // rectangle and silently skip pixels that are actually on the surface.
+        int x0 = (int)Math.Max(0L, (long)cx - outerRadius);
+        int y0 = (int)Math.Max(0L, (long)cy - outerRadius);
+        int x1 = (int)Math.Min(Width - 1L, (long)cx + outerRadius);
+        int y1 = (int)Math.Min(Height - 1L, (long)cy + outerRadius);
 
         for (int y = y0; y <= y1; y++)
         {
